Checkpoint: Made bookkeeper files
